network: limit resolvers to virtual router srv IPs
This was found as we noticed that the static old specific IPs in DEV
caused telemetry outages as nginx would try to resolve using the dead
IPs.

Those IPs are now independent of any specific hardware life cycle
and consistently use SRV which is available to all nodes directly.

Fixes PL-132394