router: only enable BFD in Bird on transfer interfaces in DEV
Try to reduce the degree to which Bird and FRR step on each other's
toes when using BFD. FRR sets up BFD sessions with the top-of-rack
switches, and Bird sets up sessions with the uplink routers. However,
due to overlapping socket bindings, Bird sometimes receives packets
intend for FRR, and prints a confused message in the journal every
time it receives an errant packet.

Hence, restrict the interfaces on which Bird listens for BFD to reduce
the amount of log noise.