php: use glibc 2.38-66 (temporary hotfix)
This vendors the glibc expression from upstream nixpkgs, as it is still in staging there.
To avoid having to rebuild all depending packages on our own, we only
rebuild php with the new glibc for now as there are teasers that the
highest exploitability of the CVE is through php.