firewall: don't log refusals to console
To further reduce the impact of network attacks have on serial logging,
we do not print the firewall refusals to the console anymore.
As we still want to see the regular systemd messages on the console that
are logged with `6`, firewall log level needs to be increased to `7` and
`consoleLogLevel` is cut off at `6`.

PL-132397