networking: reduce connection tracking usage
We have a lot of connections (Ceph, VXLAN) that are allowed in the
firewall without having to keep track of state but will have a high
number of flows/combinations associated.

To avoid the connection tracking table to explode, we exclude Ceph
on the client and replication network, as well as the VXLAN underlay
traffic.