ceph/mon: fix service start after boot
fixup for c534b9bd1c3e05b45a05a50d6f8b154147403f18

In https://github.com/flyingcircusio/fc-nixos/pull/995 we added a start precondition of a mountpoint existing to fc-ceph-mon.service:

`ConditionPathIsMountPoint = "/srv/ceph/mon/ceph-${config.networking.hostName}";`

Unfortunately, the actual mounting of the mon volume to that path is done by exactly that service that is prevented to start by that condition.

The main reason to introduce this condition was just to make the tests less noisy when a test VM started for the first time, without having a mon volume created yet. We can simply remove that condition at the cost of having more red herrings in the test output.