router/firewall: reopen port 80/443 as we need Let's Encrypt.
Those ports were turned off during the router migration to NixOS
because we thought they were only there for historical reasons when
we used to present awstats on this port. Nowadays, we do leverage
let's encrypt for internal names in many places and this breaks
certificate generation/renewal.

VMs protect themselves on 80/443 anyway, so this is safe.

Re PL-132680