mongodb: add to important packages
To ensure that mongodb binaries are available in our cache, readilyl
availble for customers to build their own deployments without having to
compile the package on the VM, the package is marked as important.

`mongodb-6_0` because this is the version currently interesting, and
`mongodb` to see when there's a major version change in the package
version diff.

FC-38101