docs/mysql: clarify root and super user
In the docs, the terms "super user" and "root" had been used interchangeably. To clarify that both terms always refer to *root*, I decided to unify this.
Background: This confused me personally a bit, as with some other DBs and socket auth this disticntion can be a meaningful one. But here, socket auth is not used anyways.