monitoring/network: check explicit list of interfaces
We used to "autodetect" the interfaces to check. However, with recent changes there may be intermediate states where interfaces are stuck in a semi-configured state until reboots.

This change adapts the interface check to only verify the explicitly named interfaces that we expect to be up. We don't really care about others.

Re PL-132536