fc-luks: `test-open` subcommand for testing unlock
For streamlining the regular task of verifying whether (all) volumes of
a host can still be opened with the admin key and device key, I
implemented the `fc-luks keystore test-open` subcommand.

As this relies on the same volume discovery logic as the `reencrypt`
command, I refactored and reused that discoverly logic out into a
separate method.

Also includes a NixOS integration test for this, but no unit tests.