physical machines: sensu check that swap is off
When doing full-disk encryption, persisted encrypted data from disk can
still be kept in memory in a decrypted state. When swapping memory pages
to disk again, there is the danger of persisting that plaintext data to
disk again, circumventing the goal of the encryption.
As we do not want to use swap on physical machines anymore, we can just
monitor and warn when swap is unexpectedly used.

PL-131325