restore-single-files: reduce chance of UUID collision
XFS can't mount multiple images that happen to have the same UUID.
This can happen either due to PL-132849 or if you'd like to restore
from multiple revisions of the same disk.

This change reduces the chance by changing the UUID during restore.
However, there's a small mount/umount cycle with the old UUID that
could collide, but xfs needs a clean log to change the UUID.

Re PL-132849