Allow jitsi-meet which is marked as insecure now
This is caused by libolm used for e2ee which is an optional feature.

There's no fix in sight (libolm deprecated, no signs in lib-jitsi-meet
to move away from it) and the attacks are AFAIK for on the theoretical
side. I don't think that this should stop us from using Jitsi.

PL-132999