webproxy: fix varnish state dir check when reloading
`/run/varnish/<hostname>` can be an empty dir and the script would just
check existence as a dir dir when reloading, before this change. An
empty dir causes the reload to fail because varnish cannot find its PID file.

Checking for the PID file directly should fix the problem (untested!)