varnish: perform VCL syntax checks at runtime
Re-enable the upstream syntax check and perform syntax checks on
all vhost configs.

This might cause unexpected build time failures that are currently
hidden if users include external files, but as discussed with
AppOps - if that is the case and we disabled the checks previously,
then that's where we were wrong and need to fix the deployments.

Re PL-132278