fc-ceph-load-vm-images: rework image cleanup
In the ideal case, only the single current snapshot of an image is kept.
In case of a race condition, we try again at the next run.

Always keep at least one snapshot as a safety measure against invalid
image list data.

suggestion how to make the code a bit easier to read