kernel: pin stable version to 5.15.164
We are trying out a 6.11 kernel in non-prod right now, for (most) prod
systems we like to keep the 5.15.164 kernel we have been using for some
time now. Before, we used a revert in our nixpkgs fork to get the
desired version but we have multiple kernel updates from upstream
nixpkgs now and we want to pin it here to avoid any unwanted
updates and confusion.

PL-133043