nfs_rg_client: remove reload service
After switching to the `hard` mount option in PL-133015, retrying is
already done by the mount intself and does not need to be imposed by a
wrapper service.
Even worse, in PL-133062 we've seen that the retry logic (in that case
when reloading) can turn into an infinite loop that blocks a whole
system switch. Remounting an NFS mount is brittle anyways, as changing
of most of the mount options is not supported anyways.