nfs roles: unify hostnames to match /etc/hosts
We record all other hosts of the RG in the system’s /etc/hosts to enable
hostname lookups even without an active DNS resolver.
But as the directory currently exports these hosts with a different
domain (gocept.net) than what we record in /etc/hosts
(config.networking.domain -> fcio.net), this lookup path does not work.
This can break setups using a local DNS resolver, e.g. openvpn with
dnsmasq.

As a specific workaround for the nfs role, we normalise the hostnames
used by NFS to match the format we expect in /etc/hosts.

PL-133063