agent: trigger cold reboots when KVM environment parameters change
Add support for a mechanism in fc.qemu similar to the existing qemu
binary generation handling which additionally seeds various parameters
related to the KVM environment at boot time and which writes these
parameters into /run periodically at runtime. Request a cold reboot
when the KVM environment parameters change at runtime to allow the
hypervisor to perform changes which require the qemu process to be
offline or restarted.

PL-131857