provide uids for the provisioned statshost datasources
All statshosts will now share the same uid for their datasources
explicitly so that dashboards can be provisioned easier.

The uid for the prometheus datasource was taken from the dashboards in
the grafana respistory, the uid for the loki datasource is newly
generated for use in a new dashboard.

It appears that the prometheus data source uids were identical across
statshost probably due to how they are generated in grafana, however
being more explicit helps streamline creating new dashboard provisions.