s3users: introduce an additional safety-belt
We decided to simplify the directory API as we do not have a specific
use case to let storage servers discover S3 users from other clusters.

This is due to the caution to not accidentally end up deleting data
for a user in the local cluster just because someone wanted to delete
a user in a foreign cluster and due to simple bugs we end up seeing this
in the wrong cluster and act on it.