python38*: work around eval failures
Eval was failing due to recursing into all python packages, trying to
eval them with python38. But some of them are explicitly marked to not
build with that python version, throwing an eval failure.

We don't need to recurse into the package set with our release logic,
because the packages important to us are mentioned somewhere else
explicitly and then pull in the specific python packages relevant to
them.