tests.router: fix system path initialisation
The `Machine.script` property is not accessible from the python test
runner anymore. It was used to retrieve the top-level system nix store
path without any specialisation being active.
The approach was already a bit cursed, as it involved regex-parsing the
test script for the kernel path.

Instead, a similarly cursed approach generates python statements from
Nix code that initialise the python instances with appropriate values
derived from the Nix config.

The tests still fail and automatic specialisation switching does not
work, but I verified that the store paths used by the python code are
the correct ones.