agent: add explicit subnet id generation to kea config
As of Kea 2.6.0, automatic generation and assignment of subnet
identifiers is no longer supported. Our Kea config generation script
relied on the automatic assignment so we didn't have to do that
ourselves, however this point is now moot.

This uses a counter to assign an index to each IP subnet as the config
is generated. Adding or removing subnets from the directory will cause
the IDs to change and potentially invalidate the lease database files,
but given how rarely we change our IP space and that we mostly have
DHCP service in order to hand out static leases, this isn't a problem
worth worrying about.