router: use vtysh for testing zebra liveness
FRR now uses a single systemd unit and the watchfrr watchdog program
for managing the FRR daemons, instead of having a separate unit for
each daemon. We can't use systemd to test whether zebra is running
directly, however we can use vtysh to connect to the zebra CLI socket
as an alternative.

PL-132122