sensu check network_speed: only enable on virtual machines
The network_speed sensu check was introduced to catch regressions and
performance degradation in VM netorking drivers.
As it relies on the rgw.local S3 gateway provided byt he VM's KVM host,
the check fails on physical machines. As we do not strictly require such
a check there, just disable it.

PL-133319