[PL-133028] reorganize Grafana dashboards in folders based on the source's folder structure
The dashboards sourced from the Grafana dashboard repository are organized in file system folders.
This structure can be re-used for provisioned dashboards in the Grafana UI for an improved separation
between "Core" FCIO Dashboards and "Feature Lab" dashboards that can serve as the basis for custom
Dashboards in individual Grafana instances.