fc-slurm check: don't crash on non-"slurm-node"s
When executing `fc-slurm check` on a host that does not have the
`slurm-node` or `slurm-controller` role, but is e.g. a
`slurm-external-dependency`, the `fc-slurm check` command used to crash
due to no available node data.

This implements a fallback warning message to better understand the
situation and, if this was about to happen on an actual slurm cluster
node, better debug the available cluster information.