tests: use driver-assigned id number for mac addresses
The NixOS test driver assigns each test VM an ID number which is used
for generating MAC addresses. As this number might diverge from the ID
we provide in our test framework for e.g. assigning IP addresses, we
should use the driver-provided ID for MAC addresses so that interfaces
are detected correctly.

PL-133360