rabbitmq: fix unit file generation
optionalString generates an empty string if the condition
is not met. This resulted in an overides.conf for rabbitmq.service which just contains some auto-generated env vars in [service].

For the "frozen" rabbitmq 3.6 service/role, the new PATH replaced the PATH set by the original unit, leading to a "command not found" for rabbitmqctl in `postStart`.

Setting feature flags isn't possible for rabbitmq 3.6, the rabbitmqctl command must only
run for current rabbitmq versions.

For other machines *without* rabbitmq, this also generated an empty unit file because there was no condition to check if rabbitmq is actually enabled on this machine.

We need to limit this to the current Rabbitmq role and use mkIf here to prevent setting the option at all when
conditions are not met.

PL-133374