nixos/platform: set Nix's connect-timeout to 1s
PL-133327

The overall idea is to no longer depend on Hydra itself for management
commands and deployment pipelines to be functional: Hydra will be served
under a new domain soon and `hydra.flyingcircus.io` will only provide the
binary cache and the download URLs for build products such as channel
tarballs. Otherwise we'd have to fix up an unknown number of Nix
expressions on all of our VMs.

The download URLs and binary cache are served by two VMs that proxy
these requests into S3: build products will be synced into S3, the
binary cache is already S3-based. To make sure this setup is still
available when the VM gets into maintenance, we'll use two VMs and
`hydra.flyingcircus.io` will have records for both of them, i.e.
DNS round-robin.

For this to work properly, we have to set `connect-timeout`: otherwise,
Nix would try to connect to the IP of the VM in maintenance for a long
time before trying the other IP. After a few substitutions, Nix would
retry the unavailable IP again. The `connect-timeout` is there to nudge
Nix to try the other IP after a very short amount of time.

The value of 1s is chosen for the following reason: the machines are
running in a datacenter and the ping between rzob and whq (where the
binary cache is hosted) is roughly ~10ms, so a timeout of a second
is pretty much. Nix retries substituters five times (see option
`download-attempts` in `nix.conf(5)`), so if the timeout is reached
regardless, there will be up to five retries before giving up.

Unfortunately it's not possible to set this per substituter. However,
cache.nixos.org is hosted on a CDN (fastly) and the ping is still pretty
low (~4ms from rzob, ~7ms from whq).

(cherry picked from commit a8680e283cb64696f762581a95dd3e030a2a5d49)