nixos/platform: use `nix.settings` for connect-timeout
PL-133327

As discussed, it might be useful to have some kind of feature-flag for
the connect-timeout since modifying declarations of `nix.extraOptions`
is hard.

Instead, decided to just use `nix.settings` here, that way the timeout
can be adjusted with primitives of the module system and without an
additional option.