fc-devhost: more resilient cleanup after disk image failures
This makes the cleanup phase for dealing with image mapping failures
more resilient:
If an image mapped successfully but was never mounted, the cleanup code
did not unmap it because the unmount had failed.
Better try to execute all the cleanup steps anyways and just report
about individual failures aftrwards.