roles/kvm: fix option description eval
`literalExpression` was not taken from lib and hence caused the
fc-search eval to fail.

I also checked all other occurances of `literalExpressions` in the
platform code. Almost all of them ar placed after a global `with lib;`
which is bad style but at least causes them to eval fine.

PL-133423