roles/lamp: make interpreters of all pools available globally
FC-42506

It's a common pattern in a lot of projects to expose `php-{name}` (or
`{name}-php`) which wraps the interpreter used by the phpfpm pool.

This patch adds this behavior to the platform and updates the docs
accordingly.

Existing projects are still able to upgrade with `php-{name}` in
`systemPackages` since the priority of each wrapper package is set to
low.