monitoring: make port speed monitoring more flexible (and sane)
We used to hard code that anything between 1G and 10G is fine.

However, we want some networks (sto/stb in the old world, the underlays
in the new world) to always be at least 10G.

However, mgm, fe, srv are fine with 1G.

And last, but not least: we don't want to be alerted if a port is
faster. Technology makes progress and we just got critically alerted
because a link was now at 25G ... scary!

Fixes PL-133472