routers: fix pmacctd traffic accounting bind to interface
Fix a regression in pmacctd after it was updated from 1.7.5 (NixOS
21.05) to 24.11 (1.7.9), where the `interface` config key was removed.
The same functionality can now be configured with `pcap_interface`.
The change can be narrowed down to pmacct-1.7.7, revision
8583beaadc4b8d559b84edc0f7273e59aea842d8, but was never properly
documented by the upstream project.

This had the effect that pmacctd ignored that config option and fell
back to accounting *all* interfaces. As a result, traffic was accounted 3
times (incoming physical, vx… bridge).
As we are running 2 instances of this service, originally designated
to account only fe and srv each, we get 2*3 = 6x.

Additionally to updating the config key name, our NixOS test now checks
for any warning issued by pmacctd, and also verifies that log output
confirms binding to the correct interface.