tests.router: mock away fc-trafficclient.service due to missing directory
The fc-trafficclient.service tries to read directory connection
parameters from /etc/nixos/enc.json and then connect to a directory.
This fails in the test, causing the whole service to fail.

In the router.maintenance test, this caused the system activation to
fail when switching back froum router2 to router1 as primary.

The test was already quite flaky for some time, but now looking at this
issue I honestly do not know how it had managed to succeed at leas
*sometimes* before.