nixos/platform: change default nix to 2.18 (prod) or 2.25 (non-prod)
PL-133484

Nix 2.24's inefficient Git tarball cache is a major problem for VMs[1] with
low IOPS (default is 250). The way forward is

* switch back to Nix 2.18 on all production VMs.
* roll out Nix 2.25 as designated successor to all non-production VMs
  (approach derived from the verification kernel topic from
  eba8be36bd4c5490dbf23ed9265f44fea92e3e42).

The challenge here is that `flyingcircus.agent.package` is supposed to
be modified by downstream consumers, e.g. for Slurm support. Hence, the
option isn't modified directly, but each agent package is post-processed
via `apply` to use `config.nix.package`.

A nice side-effect of this is that setting `nix.package` also changes the
Nix used by the agent, so `nix.package` behaves as I'd expect it. Please
note that this also means that setting `nix.package` to e.g. Nix 2.26
implies a rebuild of the agent now.

I decided against overriding `pkgs.nix` with an overlay since there are
a bunch of packages out there that explicitly require a specific Nix
version, so the potential fallout from that is higher than modifying
`nix.package`.

Additionally I changed the usage of Nix 2.24 in the following places:

* The PATH of `fc-collect-garbage.service` doesn't have a Nix at all
  anymore. The agent package already prefers its own Nix, so this had no
  effect at all.

* The agent isn't built against 2.24 in pkgs/fc: I see no reason to do
  that since there's zero usage of this. It's now built with Nix 2.18
  since that's what the majority of all VMs will use for now.

  The variant with 2.25 is also built by Hydra because of the VM-test,
  so staging VMs don't have to build their own agent.

* The sensu check-env uses the default Nix as well.

[1] See PL-133484 for measurements with Nix versions and related
    upstream bugs.