add tracing for when evalution recurses into 21.05 nixpkgs instance
When the evaluation traverses into the nixpkgs 21.05 revision we still
ship for certain hardware and infrastructure features, issue a trace
message. This helps debugging the eval depth and discover cases where
that path is evaluated unexpectedly, we don't expect that to be the case
on virtual machines.

part of PL-133522