rabbitmq-server: 3.12.13 -> 3.13.7
Necessary to enable all **stable** feature flags before the upgrade to
version 4.x in NixOS 24.11. In 3.12.x, the required flag
`stream_filtering` was not stable yet, hence not automatically enabled.

The only changes pulled in are the update of rabbitmq-server and a NixOS
adaptation for the used Erlang version.

PL-131360