tests: disable automatic ip address assignment by the test harness
By default, the NixOS test harness will automatically assign IP
addresses to interfaces inside the test VMs based ont he VM's internal
ID number and the index of the VLAN the interface is connected to.

However, if we're building a test VM using testlib.fcConfig, then
we'll be providing ENC data with IP address configuration which will
be consumed by the platform to set IP adresses itself. Let's ask the
NixOS harness not to do this, we should only be configuring IP
addresses in one place.