nix: use Nix 2.25 on all machines
So far, 2.24 is still the default Nix version in nixpkgs, but we know
that we do not want that one. So use the opportunity of a development
cycle to use a more modern nix everywhere.

Keeping the test, as we very likely want to try out newer Nix versions
in a location- and nonproduction-specific way in the future again.