nginx: make HTTPS sensu check work for vhost name != serverName
The assumption that server names are always determined by the attribute
name of the vhost was wrong. The check now prefers to talk to the host
identified by vhost.serverName, falling back to the attribute name.

Now, we also check if a vhost really uses SSL as there might be
non-TLS but ACME-enabled vhosts which just serve the ACME challenge.

PL-131278