kvm: provider resolver on virtual ipv4 gateway using dnat
Guests with layer 3 routed interfaces may not necessarily be managed
by the platform, so we may not have the ability to update the network
configuration in guests after it's been baked in during
bootstrapping. The DNS resolver address configured is one such item,
which means that we would have problems if we ever wanted to change
the resolver address if there were unmanaged VMs using that resolver
address in the same location.

Instead, we can let the guests use the virtual IPv4 gateway address as
their system resolver address, and then use some DNAT rules in the
forwarding firewall for guest interfaces in VRFs to transparently
forward requests to the site resolver. If the site resolver address
then ever changes in the future, we only need to update the firewall
on the KVM hosts instead of breaking in to every unmanaged VM which
relies on the old resolver address.

PL-133324