docker: enable forwarding in role
The docker role requires IP forwarding to be enabled so that
containers are able to access services hosted in the outside
world. The upstream NixOS module enables forwarding when docker is
enabled, however the platform disables this again by default. Add an
override to the general platform override to make our local
configuration match that of upstream.

Set the override priorities manually in order to avoid evaluation
conflicts where external_net is also enabled.

PL-133589