external_net: configure forwarding to coexist with the docker role
As both external_net and docker roles may be configured on the same
machine, this can cause evaluation failures if both roles configure IP
forwarding with the same priority. Make this interaction explicit with
a note and a numeric override priority instead of using an opaque
utility function.

Additionally, on modern Linux net.ipv4.ip_forward acts as an alias for
net.ipv4.conf.all.forwarding. The latter is the name we use in other
roles, so add it to external_net (along with the setting in the
"default" namespace which we also set in other roles) for consistency.