hardware: configure sysctls for accepting connections in vrfs
By default, services bound to interfaces outside VRFs are not
accessible inside VRFs and vice versa, and services bound to an
interface in one VRF are not accessible from another. This means, for
example, that VMs whose interfaces are bound to a VRF are not able to
access e.g. SSH or Telegraf running on the KVM server in spite of the
KVM server forwarding guest traffic inside the VRF.

However, on routers we want to provide some services to all connected
networks whether VRF or otherwise (in particular, the DNS
resolver). Enable the relevant sysctls in the router role in order to
allow locally bound services to be reachable to connections received
in VRFs.

PL-133324