tests: test for convergent forwarding settings in docker test
If the IP forwarding sysctls in the system configuration are set to
values which differ from the settings which the Docker daemon sets
when it starts then this can cause forwarding to be turned off again
if the system sysctl config is reapplied during a config switch.

Our existing test doesn't catch this problem because it only tests the
sysctls once after Docker has started up, so extend the test to check
that reloading the system configuration does not diverge from what
Docker expects.

PL-133589