fc.ceph: correctly specify python dependencies
Eval failure since NixOS/nixpkgs#394838

The expression is called via pythonPackages.callPackage, which takes
care of populating the dependency function args with the packages of the
correct python version.
Hence, python packages shall not be taken from a passed
`pythonXYPackages` set but be specified directly in the function args
and used from there.