roles/webproxy: filter varnish listen addresses by uniqueness
PL-133646

About a year ago, the devhost configuration only added `[::]` to the
listen addresses, however this only added an IPv6 listen breaking
certain deployments on devhosts if they explicitly talked to Varnish
with IPv4[1].

As a band-aid fix I added the `0.0.0.0` http_addr to said deployment in
the devhost configuration which worked until I upgraded fc-nixos to a
version with the fix.

While this is something to be cleaned up in the deployment, it's also
wrong to generate a broken configuration here, so making the list unique
still seems reasonable.

[1] 70bc3858ba0365e1b9563c16a3a9c28e2dbef557