tests.statshost-global: serve stub metrics via actual telegraf
Example stubmetrics used to be served by a python http-server instance
on statssource. But as prometheus>=3 becomes more strict in the
`Content-Type` header it accepts, the default
`Content-Type: application/octet-stream` of the python server became an
issue.

Instead, we know inject metrics into an actual telegram via the socket
input. This also tests the compatibility of telegraf and prometheus.