fc.agent: improve test coverage for `release_path`
Provide test coverage for the code path responsible for PL-133713.
Also corrects some type annotations.

The switch from `os.readlink` to `os.path.realpath` was done to simplify
testing: `readlink` *requires* the target to be a link, while `realpath`
only *allows* it to be. We do not have to enforce system paths to be
links though, this is more an implementation decision of NixOS.