services/redis: stricter naming rules for redis servers
Follow-up for PL-133583 PL-133563

I realized that using dashes in the the Redis server-name and a password
causes the envsubst script in Telegraf to fail because it injects the
server name into a variable.

I consider this to be pretty unlikely though: we haven't documented
anything on multiple servers, so most uses probably stem from nixpkgs
itself and in that case, the server is usually local and thus
authentication is done without a password.

This patch restricts the server-name to alphanumeric chars, underscores
and dashes. Dashes are escaped with two underscores (__). Using a
server-name with two underscores already is prohibited.