[PL-133737] alloy: add supplementary nginx group only if nginx enabled
Move the `lib.mkIf` up one level to include the second entry in the
`lib.mkMerge` block which adds the supplementary "nginx" group to the
alloy service which allows alloy to read the json-formatted nginx access
log.

This fixes an issue with the alloy service being assigned the nginx role
even when nginx is not enabled.