tests: integrate interactive SSH debugging into our test suite
PL-133751

With that, it's possible to do

    nix-build tests -A redis.driverInteractive

and use the ipython-based repl to step through the test. For tests from
fc-nixos, the SSH backdoor is enabled by default, i.e. you get
instructions on how to log into the VM.

This also reworks the test setup a bit: I removed the recursive test
discovery since this removes `driverInteractive` among other useful
attributes.