tests.ceph-nautilus: introduce retry for some flaky actions
The osd rebuild on host2 was failing some times due to the OSD being not
ready to be stopped again, despite the cluster having been clear before.
As a workaround, wrap this in retry logic.

This is only intended as a workaround pattern until the tests are
re-architected together with the next ceph update. Until now, new
failures might be worked around by introducing `retry_attempts` as well
for the particular line.