mailserver: migrate policyd-spf -> rspamd
Upstream simple-nixos-mailserver has removed policyd-spf support, so our
SPF-related options need to be migrated to rspamd as well.

This renames the option to a more generic name, and uses it to populate
the rspamd whitelist of the spf module (enabled by default).
policydSPFExtraSkipAddresses -> spfSkipAddresses

The hardcoded localhost `skipped` addresses of the previous option are
already part of the general global rspamd `local_addrs` option.

The only thing I am not fully certain about is whether the previously
excluded hostnames from `HELO_Whitelist` are now excluded by rspamd as
well.

PL-133285