ceph-load-vm-images: only lock image if there is no lock yet
We've seen multiple times that locking an image works but unlocking doesn't
due to DNS issues. This lead to a locking failure in the next run.

To fix this, use a general lock cookie for the update images script and
only lock when there is no lock on the image yet.
Concurrency protection is handled by systemd.

PL-133726