storage tuning: change to static memory sizes for dirty page buffers
We discovered that static buffers will provide more consistent latency.
In a number of situations the buffers grew very large on systems with
lots of RAM and that seems to have caused IO starvation/blocking
resulting in applications becoming unresponsive.

After some thought and discussions we figured that conservative dirty
buffers are better suited in the general case.

Re PL-133760