roles/ceph/rgw: add s3/object gateway on rgw machines
PL-128135

We no longer let radosgw listen on port 80, but on port 7480 instead and
let the reverse proxy in front listen on port 80.

The haproxy in between uses other rgws belonging to the same cluster as
backup servers.

TLS termination is done with nginx. The ACME challenge is DNS-01 since
we usually have multiple RGW instances.