agent: Unify IO weight fallbacks with VM overrides
`IODeviceWeight` as well as `IOWeight` both affect the cgroup's
`io.weight` property. An IODeviceWeight is just seen as more specific
and overrides that weight for a certain disk[^1], if present.

Hence, the bare `IOWeight` will be only relevant on non-VMs (and
database VMs). These fallback values should then be at least in-line
with the device-sepcific overrides, making them smaller and more
restrictive is unnecessary.

[^1]:https://docs.kernel.org/admin-guide/cgroup-v2.html#io-interface-files