devhost: simplify upgrade to platform 25.05
type="vm" is the only supported value here anyways for this release, but
having the option default on "container" requires a manual override on
the machine to "vm".
Having removed this option altogether in 25.05, this creates the need
for a config that does not build on both releases and requires a manual
config change at upgrade.

Changing this default here allows removing the local config override and
simplifies the update process.