roles/jitsi: don't enroll coturn secret at eval-time
PL-133672

Generating the secret for prosody/coturn at eval-time via IFD has the
problem that the derivation is not in the system's closure and gets GCed
eventually. In rare cases this lead to restarts of Jitsi mid-day because
of a GC and a subsequent rebuild of the password.

It's a little hard to test this correctly because Jitsi seems pretty
resilient to a broken coturn. However, with the password-file not
existing for coturn, the service failed to start up. With the password
mismatching I got the following errors in coturn's log on a test
video-conference with two devices:

    Jul 25 18:37:17 test42 turnserver[617432]: 17: (617438): ERROR: session 002000000000000002: check_stun_auth: Cannot find credentials of user <1753547825>

This only happens on a secret mismatch.

Prosody's Lua code would also fail hard if it couldn't read the secret:

    out>    Jul 25 19:08:22 test42 prosodyctl[634356]: A problem occurred while reading the config file /etc/prosody/prosody.cfg.lua
    out>    Jul 25 19:08:22 test42 prosodyctl[634356]: Error: /etc/prosody/prosody.cfg.lua:97: /run/credentials/prosody.service/turncredentials_secret2: No such file or directory