tests/ceph-nautilus: add testcases for rados/s3
PL-128135

I think that the coverage for S3-like features is too low in this
codebase. This patch confirms the most basic operations (mb,
up/download) and checks that presigned URLs work as expected.

This depends on a port-forward from port 80 to port 7480. Given there's
more work pending - a gateway should be added to rgw instances - I'd
like to make sure we have tests for this to reduce the likelyhood of
regressing.