fc.agent UpdateActivity: drop unused changelog and release name
As reflected in the test data, the maintenance announcements once used
to include the Flying Circus platform release names (like 2021_013), as
well as changelog URLs.
Checking the current behaviour on live systems, this is not the case
anymore. This is no wonder because there is no current code path to
initialise these attributes:
- the data needs to be read from ENC for serialisation
- the code for that, enc.py `get_release_info()`, is never used
- so the attributes can only be set by old serialised maintenance yaml
  data. All of these are gone now, as we do not serialise any new ones.

  For now I keep `get_release_info()` around in case we ever want to add
  back changelog URIs to summaries, e.g. in PL-134037. But the rest of the code path was
  dead, potentially misleading code.