fc.agent: UpdateActivity: don't serialise `current` attributes
While the target state of a NixOS update activity is well-defined, the
update always runs based on the current system state. In case that state
has been modified between time of first serialisation of that activity
and the time of execution, the `current_`* attributes are stale and
outdated. It does not make sense to serialise these attributes for later
use then.

With regard to decisions based on that *current* state of maintenance
scheduling, like required kernel reboots or restarting units, this can
of course change when the `current_` attribute changes.
But what we had done and announced so far was already at the danger of
getting stale and incorrect anyways. Checking what was announced and
what is now actually going to be executed is left as a future
improvement.

(cherry picked from commit ae7088b562c10ed2c30aa170c4d3d150a9c2bc57)