hardware: move vfs cache settings into backy option
The settings used hitherto for VFS cache eviction are no longer
suitable as general configuration for all hardware
profiles. Aggressive filesystem caching is not necessary for
filesystems on SSD-backed storage, and we've observed cases where
processes with high memory usage get OOM-killed because the kernel is
too reluctant to evict the VFS cache.

Move the sysctl behind a module option in the backyserver role to keep
it available on older HDD-based backup servers.

PL-133712