fc-agent: compare kernel packages instead of versions
For checking whether a reboot is required, we compare kernels of current
and next system. The previous logic tried to extract the *kernel
version* from the store by investigating lib module paths.
This was brittle and suffered from multiple issues:
- we encountered a system where the expected subdirectory tree structure
  was not present -> causing a crash
- kernel changes within the same version, e.g. recompiling with
  different config, are not detected
- so we actually do not need to compare the kernel *version*, but exact
  *package*

This introduces a KernelIdentifier denominating the used kernel
packages. It deliberately can only be compared for equality, and
provides some robust pretty printing of the identifier itself.

fixes PL-134082